Custom Train/Test Ratio
The test_basic.py file works once again; apparently, subprocess can hang if stdout is piped but not regularly flushed.
The solution is to send the subprocess stdout straight into a log file.
This commit additionally allows a streamed .csv file to be broadcast at custom ratios by a data server and a query server.