Rudimentary Data Type Inference
There is now a structure in DataStorage that associates lists with data types.
A basic inference using ast.literal_eval() is enabled; the data types are determined by the first element that gets inserted into a list.
However, a lot of TODOs remain, such as (1) making inference more robust, e.g. to NoneTypes, (2) considering what to do if an element arrives with different data type, while acknowledging the difference between regular and anomalous values, and (3) allowing merged lists to generalise data types.
The design would also be better if the DataPorts handled the type inference/conversion rather than DataStorage.
However, for now, it's time to progress to model development.