Few Asyncio Fixes
When an asyncio task is constructed from a bound method, it maintains a strong reference to self. This means that objects that have otherwise dropped out of scope are kept alive by infinitely looping asyncio tasks.
This commit introduces a new function that converts a bound method into a weakly referencing task that terminates itself once the binding instance is out of scope.
This function still has to be updated to allow for additional args. The task that serves as an initial entry point also has to be reconsidered, as it strongly references.