First Hyperparameter and Linear Regression Conundrum
Ahead of multiprocessing, this commit establishes, among certain improvements, an initial hyperparameter scheme.
So far, the only hyperparameter is a custom batch_size that determines whether River components will operate in batch or incremental mode, each designed for their own data format.
As a current conundrum to understand/solve, River's predictors appear terrible for the SPS problem, with or without a standard scaler.