Dedicated Asyncio Event Loop
The asyncio.run() function is blocking, which prevents Python from running synchronous code after any eternal loops are entered.
The solution is to start a new forever-running event loop in another thread, assigning tasks to that loop via a custom function.
This handle to this loop is global, like the logger, and it exists in a new concurrency.py file, along with custom methods needed to schedule tasks into this loop.
The lifetime of the thread/loop is eternal and left up to garbage collection when a kernel closes; whether this is appropriate is an open question.
Nonetheless, the benefit is that asynchronous AutonoML operations now work identically in Python/IPython mode, recalling that the latter introduces its own main-thread event loop.