Need For More Asynchronous Tasks
The problem with running an asyncio event loop on a non-primary thread means that all synchronous code on the main thread can clash in arbitrary order.
User-called functions, as a basis, need to schedule their processes in the event loop rather than run them in parallel. This is the intention for the next commit.