Main Thread Plots
This is a quick checkpoint with a few minor improvements of the code, including getting plots to work. Essentially, matplotlib doesn't seem to plot figures in a secondary thread, so pipeline plot figures have to be constructed and then passed back up to the synchronous UI level before being displayed.
In any case, this is also a checkpoint, with multiprocessing still not working. The asyncio package has a ProcessPoolExecutor, so that is the next idea to investigate.