Minor Checkpoint for Component Subfolder
It is now possible to dynamically import modules within the components subfolder of autonoml, where sklearn and river components have been shifted.
This includes generating dicts of preprocessors and predictors that are not tagged with 'is_unselectable = True' and are either at the maximum depth of their MLComponent subclass hierarchy or have all their children tagged with 'is_unselectable = True'.
However, there is a multiprocessing pickling problem of a dummy regressor imported elsewhere, seemingly because the strategy.py file is reimporting modules (or vice versa).
Needs to be solved.