Champion Challenger
A ProblemSolution class has been introduced that, at the moment, stores a list of pipelines. Whenever a new one gets inserted, the one with lowest loss gets discarded. In effect, this allows a champion-challenger paradigm to be run.
HPO run information is also sent back to the main process now. Additionally, creating supportive HPO workers is now properly awaited so that, if the HPO name server closes before they can join, the exceptions are caught and, if the HPO run is complete, the support is cancelled.