Refactor Data Storage
The observations and queries attributes of the DataStorage are now dictionaries of DataCollections rather than DataCollections themselves. The collection_ids are associated with tags that the user provides when ingesting .csv files (to be expanded to streams as well).
This means that allocations across these tags are now much quicker in combining DataCollections as opposed to slicing out the right samples from a superset.
The refactoring has not yet been carried out to processing queries (or strategy).