Handled Tag Superset Issue
The idea is that every combination of tags should be given its own DataCollection.
If a user marks one input file as {"a":"1"} and another as {"a":"1", "b":"2"}, they should be stored within individual partitions, even if a learner focussed on {"a":"1"} will train on their concatenation.
This commit hopefully smooths that out and also makes it easier to keep every DataCollection consistent in terms of its data keys.
Uncoupling observations and queries from the same tags might be an idea though...