HPO Semaphore
It is unclear whether this improves performance, but it is possible for the user to ensure no more than k HPO runs operate at a time, where k is specified in the strategy file and is currently set to a default of 2.
This concurrency limit may allow HPO runs to go quicker without competing for processors.