OnlineSVR Tentatively Working
OnlineSVR seems to be a slow learner and not equipped for batch learning on large data files, at least not without multiprocessing working, the efficiency of which is currently a question mark.
However, in principle, OnlineSVR does seem to work for the streaming case of test_proj_sps.py.

Pressing things to look at as of this commit include giving server reconnection attempts a limited number of retries after it has connected at least once.
Also, the order of features seems scrambled in the PyArrow representation; need to see whether this has any real impact.