Limit Reconnection Attempts
Once the data broadcasters drop out, the data receives have a limited number of retries before stopping as well.
However, this doesn't solve a problem: the servers are dropping out because they're not receiving a client confirmation regularly enough, thanks to bottlenecking on the core AutonoML thread.
One solution is to raise the delay for broadcasting servers before they shut down without confirmation; this is a settings.py value.
Alternatively, perhaps client confirmations should be sent on a separate thread/process.