Client Confirm Flexible Delay
This commit allows users of sim_stream.py to easily adjust how long a broadcaster waits for client confirmation before it abandons broadcasting.
However, this is a poor bandage on the problem.
Basically, the AutonoML machine can get choked up with operations on its AutonoML event loop, failing to send/receive in interactions with a broadcaster.
The future solution to this may require interactions with broadcasters to be based on separate processes. Alternatively, multiprocessing should be prioritised and adaptation should be passed to a loop executor.