Imputer and Scaler Pipeline Start
Short of a more flexible idea to attach components together, imputers and scalers, if included in the full search space, are now treated as their own sub-pools.
Thus, pipelines will, if possible, always take the appearance of: imputer -> scaler -> predictor
This commit verifies this works with a scaler, but explicit imputer testing is delayed for a relevant test script.