Downgrade to Python 3.11 and ConfigSpace Fix
Python 3.12 appears to have a problem with ConfigSpace, in that an internal np.round function is returning a np.float64 object that does not pass a float check.
For whatever reason, this is not an issue for Python 3.11, so, for now, the code is maintained for 3.11.
Also, it is unclear whether this issue is new, but this commit fixes a check for hyperparameters in a configspace, thus allowing for random samples to work.
With this commit, test_simple.py should demo random sampling, test_proj_iris.py should demo HPO, and test_proj_pharma.py should demo streamed learning with a default OnlineSVR.