Python 3.9+ Issue Workaround
Before Python 3.9, concurrent.futures executors could continue to take on new tasks after the main thread finishes. For 3.9 onwards, executors now refuse to take on any new tasks once the main thread ends.
This is fine for IPython, as the main thread (and its IPython asyncio loop) continues beyond the end of a script, allowing for users to interact with an AutonoMachine.
For Python, this is bad. Once the script ends, any model development processes using executors are blocked.
The workaround in this commit is a be_patient() function at the end of AutonoMachine.learn(), which is typically the last non-interactive function called.
It checks the asyncio event loop in the main thread, which should always have one task in IPython, i.e. the interactive kernel. In this case, the function does nothing and lets IPython users continue on to interact.
In Python there should be no tasks in the primary-thread event loop, so the secondary thread is waited on instead.
Currently, the second thread runs an asyncio loop forever, so perhaps the be_patient() function should simply request input instead before closing everything.
For now, this is simply a functional workaround.