Curious Online SVR Behaviour
This commit makes it possible to put pipelines in production without any validation, i.e. by selection 0 validation folds in the strategy file.
This means that streamed pipelines no longer need to wait for four data instances; they can run immediately.
In any case, despite the Online SVR query function being rewritten to be incremental, its batched predictions (once acting on multiple instances at a time) all seem to be the same, even though post-adaptation predictions vary.
It is not clear whether this is a bug or not, but it's worth noting.