One-Observation Asynchronous Adaptation Upgrade
Previously, upon a new observation encounter being acknowledged, all existing learners would be tested and adapted together on all the new observations up to that point.
This had two problems. One, observations could get clogged, with learners racing to adapt to a piling number of data points. Two, quick adapters would be held up by slow adapters.
This commit separates the testing/updating phase of adaptation so that all learners can be tested on a new observation, but adaptation requests are put on a development queue, to be handled when the time comes, and only on the last new observation. The adapted pipeline then replaces the original pipeline once adaptation is complete.