Pipeline Imports and Adaptation Fix
This commit allows pipelines to be imported from a folder when running proj.learn().
An import allocation can also be specified, details in the comments, where a filename with a substring can be allocated to a specifically tagged subset of data.
Additionally, it seems that the instruction to adapt was not propagated to pipeline development, so this has been rectified.