Adaptation Data Filtering Fix
When data that a pipeline should adapt to is filtered, it needs process info passed to it, which specified the index range to select. Without it, you can get a bad bug of constantly adapting to the very last instance in a batch.
This should not be a problem for any typical streaming scenarios. However, the setting to forcibly adapt to every instance, as used in an SPS application, throws this issue up. So, this commit fixes the problem.
Additionally, this commit allows ingested files to have custom field names.