Import Option To Do Only Allocation
There is a minor option added to the importing process.
You can now choose to only import pipelines that have a specified allocation, i.e. that have a substring in the allocation dictionary.
This allows for unnecessary pipelines to be disregarded when importing.
It is currently False by default, meaning all pipelines will be imported from an import dictionary and the allocation simply dictates what particular pipelines train on.